Fix and improve user paths through CLI wizard
