+++ tools/quickstart/quickstart/cli.py
@@ -15,26 +15,10 @@ from quickstart import bridge, docker, netstats, validator_account
     default=os.getcwd(),
 )
 def main(host_base_dir):
-    setup_as_validator = prompt_setup_as_validator()
-    if setup_as_validator:
-        validator_account.setup_interactively()
-        bridge.setup_interactively()
+    validator_account.setup_interactively()
+    bridge.setup_interactively()
     netstats.setup_interactively()
-    docker.update_and_start(host_base_dir, as_validator=setup_as_validator)
-
-
-def prompt_setup_as_validator():
-    choice = click.prompt(
-        "Do you want to setup a validator (1) a regular node (2)?",
-        type=click.Choice(("1", "2")),
-        show_choices=False,
-    )
-    if choice == "1":
-        return True
-    elif choice == "2":
-        return False
-    else:
-        assert False, "unreachable"
+    docker.update_and_start(host_base_dir)
 
 
 if __name__ == "__main__":
